raycasting block size unit and fixing meshing (0 weight => bugged rendering, fixed)

+++ VoxelEngine/src/com/grillecube/client/renderer/camera/CameraProjectiveWorld.java
@@ -43,7 +43,7 @@ public abstract class CameraProjectiveWorld extends CameraProjective implements 
 	}
 
 	@Override
-	public boolean onRaycastCoordinates(float x, float y, float z, Vector3f face) {
+	public boolean onRaycastCoordinates(int x, int y, int z, Vector3f face) {
 		if (this.world == null) {
 			return (true);
 		}
